requirements.txt
pyyaml
requests
playwright

